% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data}
\alias{add_data}
\title{Add New Zealand Regions or Electorate Type to Electorates in a Data Frame}
\usage{
add_data(data, electorateCol = "Electorate", output = c("region", "type"))
}
\arguments{
\item{data}{A data frame containing New Zealand electorate information.}

\item{electorateCol}{Name of the column in \code{data} containing electorate names.
Defaults to "Electorate".}

\item{output}{Type of additional data to append. The options are:
- "region": Adds a "Region" column with New Zealand region.
- "type": Adds an "Electorate_Type" column with electorate types.
Default is "region".}
}
\value{
A data frame identical to \code{data} but with the additional column appended.
In case of issues like unmatched electorates, the original data frame is returned
with a warning.
}
\description{
This function enriches a data frame with additional data based on New Zealand electorate names.
It dynamically adds specified data types such as regions or electorate types (i.e., "General" or "Maori")
as new columns based on the specified \code{output} type. It utilises internal datasets \code{region}
and \code{type}, which map electorates to their corresponding regions or types.

The function validates the input data frame and specified column. It converts factor
columns to character to ensure correct matching. Warnings are issued if the specified
column contains \code{NA} values, or if any electorates cannot be matched. A warning and the
original data frame are returned if none of the electorates can be matched.
}
\examples{
\dontrun{
  data(majority)
  df_region <- add_data(majority, output = "region")
  df_types <- add_data(majority, output = "type")
}
}
