% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_region.R
\name{add_region}
\alias{add_region}
\title{New Zealand Regions}
\usage{
add_region(..., column = "Electorate")
}
\arguments{
\item{...}{Data frame to have region column added to}

\item{column}{Name of column to be joined (default = "Electorate")}
}
\value{
A data frame with an additional region column.
}
\description{
Add New Zealand Regions to electorates in a data frame.
}
\examples{
add_region(majority)
}
