% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_type.R
\name{add_type}
\alias{add_type}
\title{New Zealand Electorate Types}
\usage{
add_type(..., column = "Electorate")
}
\arguments{
\item{...}{Data frame to have electorate type column added to}

\item{column}{Name of column to be joined (default = "Electorate")}
}
\value{
A data frame with an additional type column.
}
\description{
Add New Zealand electorate types to electorates in a data frame.
}
\examples{
add_type(majority)
}
