% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get NZ Election Datasets}
\usage{
get_data(...)
}
\arguments{
\item{...}{A string, naming the desired dataset.
Available datasets include:
\itemize{
\item \code{summary}: overall results at the national-level by party
\item \code{turnout}: turnout and informal vote by electorate and ballot type (candidate and party)
\item \code{vote_type}: valdidity and vote type by electorate and ballot type (candidate and party)
\item \code{party_votes}: party vote by electorate and party (electorate-level)
\item \code{candidate_votes}: candidate vote by electorate and party (electorate-level)
\item \code{majority}: winning candidate and margin of victory by electorate (electorate-level)
\item \code{split_total}: split-ticket voting (2005 - 2023 only) - the number of voters who cast
their party vote for a different party to their candidate vote (national-level by party)
\item \code{preMMP_results}: overall results between 1890 and 1993 (prior to the introduction of MMP)
(national-level by party)
\item \code{historic_turnout}: turnout between 1879 and 2023 (national-level by year)
}}
}
\value{
\code{get_data()} returns a data frame.
}
\description{
\code{get_data()} is a helper function for loading datasets. It provides the user with the ability to set
the name of the data frame within their own environment (e.g. df).
}
\examples{
df <- get_data("majority")
}
