% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{historic_turnout}
\alias{historic_turnout}
\title{Historic turnout (1879 - 2023)}
\format{
A data frame with 48 rows and 4 variables:
\describe{
\item{Election}{election year (1879--2023)}
\item{MPs}{total number of MPs in parliament (74--123)}
\item{Electors}{total number of registered electors (82,271--3.68829e+06)}
\item{Turnout}{turnout as a percentage of the number of electors (60.60--93.70)}
}
}
\source{
This dataset was sourced from the NZ Electoral Commission's official results.
The raw data are available from \url{https://www.electionresults.govt.nz/}.
}
\usage{
historic_turnout
}
\description{
The \code{historic_turnout} dataset is in a long format and can be used to look view the elector turnout at
each general election between 1879 and 2023. For an example of how to use this dataset,
refer to the Turnout article found here:
\url{https://sarahcgall.github.io/scgElectionsNZ/articles/party-and-candidate-votes}
}
\keyword{datasets}
