% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{majority}
\alias{majority}
\title{Majority (1999 - 2023)}
\format{
A data frame with 696 rows and 7 variables:
\describe{
\item{Election}{Year of election (1996--2023).}
\item{Electorate}{Electorate name (e.g., Auckland Central).}
\item{Party}{Winning candidate's party name (e.g., Green Party).}
\item{Candidate}{Winning candidate's name (e.g., SWARBRICK, Chloe Charlotte).}
\item{Votes}{Majority votes: difference in votes between first and second candidates (9--23,376).}
\item{Percentage}{Margin of victory as a percentage: majority votes / total valid votes cast (0.03--72.25).}
\item{Party_List}{Whether candidate is also on the party list (Yes, No).}
}
}
\source{
Sourced from the NZ Electoral Commission's official results, available at
\url{https://www.electionresults.govt.nz/}.
}
\usage{
majority
}
\description{
The \code{majority} dataset in a long format is used for identifying winning candidates in each electorate,
their party affiliations, and their margins of victory.
See \href{https://sarahcgall.github.io/scgElectionsNZ/articles/majority}{Majority article} for use cases.
For methodology, refer to 'Candidate Vote' section in
\href{https://sarahcgall.github.io/scgElectionsNZ/articles/nz-elections-explained.html#candidate-vote}{NZ Elections Explained}.
}
\keyword{datasets}
