% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party_votes.R
\docType{data}
\name{party_votes}
\alias{party_votes}
\title{Party Votes (1996 - 2023)}
\format{
\subsection{\code{party_vote}}{

A data frame with 12,701 rows and 4 columns:
\describe{
\item{Election}{Election year}
\item{Electorate}{Electorate}
\item{Party}{Registered party name}
\item{Votes}{Total valid party votes for the party}
@examples
party_votes
}
}
}
\source{
\url{https://www.electionresults.govt.nz/}
}
\usage{
party_votes
}
\description{
Look up the party vote in each electorate by registered party.
}
\keyword{datasets}
