% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preMMP_results.R
\docType{data}
\name{preMMP_results}
\alias{preMMP_results}
\title{Pre-MMP Results (1890 - 1993)}
\format{
\subsection{\code{preMMP_results}}{

A data frame with 162 rows and 4 columns:
\describe{
\item{Election}{Election year}
\item{Party}{Registered party name}
\item{Seats}{Number of seats won}
\item{Percentage}{Popular vote (\%)}
@examples
preMMP_results
}
}
}
\source{
\url{https://www.electionresults.govt.nz/}
}
\usage{
preMMP_results
}
\description{
View the results by party at each general election between 1890 and 1993
}
\keyword{datasets}
