% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scgElectionsNZ.R
\docType{package}
\name{scgElectionsNZ}
\alias{scgElectionsNZ}
\title{NZ Election Data}
\description{
This package makes it easy to access New Zealand's general elections data between 1996 and 2023..
The code for conversion can be found in the package directory in data-raw/
}
