% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{split_electorate}
\alias{split_electorate}
\title{Split Electorate (2005 - 2023)}
\format{
A data frame with 81,269 rows and 6 variables:
\describe{
\item{Election}{Year of election (2005--2023).}
\item{Electorate}{Electorate name (e.g., Auckland Central).}
\item{List_Party}{Party voted for in the party vote (e.g., 99 MP Party).}
\item{Electorate_Party}{Party voted for in the candidate vote (e.g., ACT Party).}
\item{Votes}{Estimated number of votes flowing from list to electorate parties.}
\item{Percentage}{Percentage of votes flowing from list to electorate parties.}
}
}
\source{
Sourced from the NZ Electoral Commission's official results, available at
\url{https://www.electionresults.govt.nz/}.
}
\usage{
split_electorate
}
\description{
The \code{split_electorate} dataset in a long format provides a detailed view of split voting at the electorate level
by party, from 2005 to 2023. Essential for understanding voter behaviour in terms of party and candidate vote divergence.
For application examples, refer to the
\href{https://sarahcgall.github.io/scgElectionsNZ/articles/split-voting}{Split Voting article}.
To grasp split voting concepts, see 'Split Voting' section in
\href{https://sarahcgall.github.io/scgElectionsNZ/articles/nz-elections-explained.html#split-voting}{NZ Elections Explained}.
}
\keyword{datasets}
