% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{split_total}
\alias{split_total}
\title{Split Total (2005 - 2023)}
\format{
A data frame with 2,435 rows and 5 variables:
\describe{
\item{Election}{Year of election (2005--2023).}
\item{List_Party}{Party voted for in the party vote (e.g., 99 MP Party).}
\item{Electorate_Party}{Party voted for in the candidate vote (e.g., ACT Party).}
\item{Votes}{Estimated number of votes flowing from list to electorate parties.}
\item{Percentage}{Percentage of votes flowing from list to electorate parties.}
}
}
\source{
Sourced from the NZ Electoral Commission's official results, available at
\url{https://www.electionresults.govt.nz/}.
}
\usage{
split_total
}
\description{
The \code{split_total} dataset in long format facilitates the analysis of overall split voting trends by party from 2005 to 2023.
Ideal for examining national-level voting patterns and the degree of voter alignment between party and candidate preferences.
For usage, see \href{https://sarahcgall.github.io/scgElectionsNZ/articles/split-voting}{Split Voting article}.
Split voting concepts detailed in 'Split Voting' section of
\href{https://sarahcgall.github.io/scgElectionsNZ/articles/nz-elections-explained.html#split-voting}{NZ Elections Explained}.
}
\keyword{datasets}
