% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{turnout}
\alias{turnout}
\title{Turnout (1996 - 2023)}
\format{
A data frame with 1,392 rows and 7 variables:
\describe{
\item{Election}{Year of election (1996--2023).}
\item{Ballot}{Type of ballot (Candidate, Party).}
\item{Electorate}{Electorate name (e.g., Auckland Central).}
\item{Electors}{Total number of registered electors.}
\item{Turnout}{Turnout as a percentage of the number of electors.}
\item{Total}{Total number of votes cast - valid and informal.}
\item{Informal}{Total number of informal votes cast.}
}
}
\source{
Sourced from the NZ Electoral Commission's official results, available at
\url{https://www.electionresults.govt.nz/}.
}
\usage{
turnout
}
\description{
The \code{turnout} dataset in a long format enables the analysis of elector turnout and informal vote by Electorate and Ballot type from 1996 to 2023.
Key for understanding voter participation and the effectiveness of electoral engagement strategies.
For application insights, visit the \href{https://sarahcgall.github.io/scgElectionsNZ/articles/turnout}{Turnout article}.
Turnout concepts explained in 'Turnout' section of
\href{https://sarahcgall.github.io/scgElectionsNZ/articles/nz-elections-explained.html#turnout}{NZ Elections Explained}.
}
\keyword{datasets}
