% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_EName.R
\name{update_EName}
\alias{update_EName}
\title{Update Electorate Names}
\usage{
update_EName(data, column1 = Electorate, column2 = Election)
}
\arguments{
\item{data}{Name of dataset}

\item{column1}{Name of column with electorates}

\item{column2}{Name of column with election years}
}
\value{
A data frame with an additional region column.
}
\description{
Comvert previous electorate names to the 2020 boundary names.
}
\examples{
add_region(majority)
}
