% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_names.R
\name{update_names}
\alias{update_names}
\title{Update Electorate or Party Names in a Dataset}
\usage{
update_names(
  data,
  nameCol = NULL,
  yearCol = "Election",
  output = c("party", "electorate")
)
}
\arguments{
\item{data}{A data frame containing election-related information.}

\item{nameCol}{The name of the column in \code{data} that contains names to be updated.
If NULL, defaults to "Party" for output "party" and "Electorate" for output "electorate".}

\item{yearCol}{The name of the column in \code{data} that contains the election years.
This column must be numeric or convertible to numeric.}

\item{output}{The type of names to be updated; either "party" or "electorate".
Defaults to "party".}
}
\value{
A data frame with updated names in the specified \code{nameCol}.
}
\description{
This function standardises names in a dataset according to the specified output type:
\itemize{
\item For "electorate" output, it updates electorate names to align with the most recent boundary changes,
facilitating consistent analysis across various election years.
\item For "party" output, it modifies party names to denote successful parties, labeling all unsuccessful
parties as "Other". This simplification aids in clear and concise visualisation and analysis.
The function is particularly useful for preparing election data for comparative and visual analysis.
}
}
\examples{
\dontrun{
  data(majority)
  updated_majority <- update_names(majority, output = "electorate")
  updated_party_votes <- update_names(party_votes, output = "party")
}
}
