% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vote_type}
\alias{vote_type}
\title{Vote Type (1996 - 2023)}
\format{
A data frame with 11,136 rows and 6 variables:
\describe{
\item{Election}{Year of election (1996--2023).}
\item{Ballot}{Type of ballot (Candidate, Party).}
\item{Electorate}{Electorate name (e.g., Auckland Central).}
\item{Validity}{Validity of the vote (Valid, Informal, Disallowed).}
\item{Method}{Method of voting (Ordinary, Special - NZ, Special - Overseas).}
\item{Votes}{Total number of votes cast.}
}
}
\source{
Sourced from the NZ Electoral Commission's official results, available at
\url{https://www.electionresults.govt.nz/}.
}
\usage{
vote_type
}
\description{
The \code{vote_type} dataset in a long format allows for the analysis of vote types (ordinary, special - NZ,
special - overseas) by their validity (valid, informal, disallowed) and Electorate and Ballot type.
Useful for understanding the distribution of different types of votes.
For usage guidance, see the \href{https://sarahcgall.github.io/scgElectionsNZ/articles/vote-type}{Vote Type article}.
Detailed explanation of voting types in 'Vote Types' section of
\href{https://sarahcgall.github.io/scgElectionsNZ/articles/nz-elections-explained.html#vote-types}{NZ Elections Explained}.
}
\keyword{datasets}
