% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{candidate_votes}
\alias{candidate_votes}
\title{Candidate votes for successful registered parties (1996 - 2023)}
\format{
A data frame with 5,357 rows and 6 variables:
\describe{
\item{Election}{election year (1996--2023)}
\item{Ballot}{ballot type (Candidate)}
\item{Electorate}{electorate name (e.g., Auckland Central)}
\item{Party}{successful registered party name (e.g., ACT Party)}
\item{Votes}{total valid candidate votes for the party (0--49,510)}
\item{Percentage}{percentage of total valid candidate votes for the party (0.00--78.06)}
}
}
\source{
This dataset was sourced from the NZ Electoral Commission's official results.
The raw data are available from \url{https://www.electionresults.govt.nz/}.
}
\usage{
candidate_votes
}
\description{
The \code{candidate_votes} dataset is in a long format and can be used to look up the candidate vote in each
electorate by successful registered party. For an example of how to use this dataset, refer
to the
\href{https://sarahcgall.github.io/scgElectionsNZ/articles/party-and-candidate-votes}{Party and Candidate Votes article}.
}
\keyword{datasets}
