% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(...)
}
\arguments{
\item{...}{Name of dataset. Available datasets include:
\code{majority}, \code{summary}, \code{seats}, \code{party_votes}, \code{split_total},}
}
\value{
A data frame.
}
\description{
\code{\link{get_data}} is a helper function for loading datasets and specifying your own name.
}
\examples{
df <- get_data("majority")
}
