% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic_turnout.R
\docType{data}
\name{historic_turnout}
\alias{historic_turnout}
\title{Historic Turnout (1879 - 2023)}
\format{
\subsection{\code{historic_turnout}}{

A data frame with 48 rows and 4 columns:
\describe{
\item{Election}{Election year}
\item{MPs}{Total number of MPs in parliament}
\item{Electors}{Total number of registered electors}
\item{Turnout}{Turnout}
@examples
historic_turnout
}
}
}
\source{
\url{https://www.electionresults.govt.nz/}
}
\usage{
historic_turnout
}
\description{
View the elector turnout at each general election between 1879 and 2023
}
\keyword{datasets}
