% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{majority}
\alias{majority}
\title{Majority (1999 - 2023)}
\format{
A data frame with 696 rows and 7 variables:
\describe{
\item{Election}{election year (1996--2023)}
\item{Electorate}{electorate name (e.g., Auckland Central)}
\item{Party}{name of the winning candidate's party (e.g., Green Party)}
\item{Candidate}{name of the winning candidate (e.g., SWARBRICK, Chloe Charlotte}
\item{Votes}{majority votes: difference in the number of votes between first and second candidate (9--23,376)}
\item{Percentage}{margin of victory as a percentage: majority votes / total valid votes cast (0.03--72.25)}
\item{Party_List}{if candidate is also on the party list (Yes, No)}
}
}
\source{
This dataset was sourced from the NZ Electoral Commission's official results.
The raw data are available from \url{https://www.electionresults.govt.nz/}.
}
\usage{
majority
}
\description{
The \code{majority} dataset is in a long format and can be used to look up the winning
candidate in each electorate, their party, and their margin of victory/majority (n & \%).
For an example of how to use this dataset, refer to the Majority article found here:
\url{https://sarahcgall.github.io/scgElectionsNZ/articles/majority}.
}
\keyword{datasets}
