% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{preMMP_results}
\alias{preMMP_results}
\title{Pre-MMP Results (1890 - 1993)}
\format{
A data frame with 162 rows and 4 variables:
\describe{
\item{Election}{election year (1890--1993)}
\item{Party}{registered party name (e.g., Liberal)}
\item{Seats}{number of seats won by party (0--67)}
\item{Percentage}{percentage of popular vote (0--58.7)}
}
}
\source{
This dataset was sourced from the NZ Electoral Commission's official results.
The raw data are available from \url{https://www.electionresults.govt.nz/}.
}
\usage{
preMMP_results
}
\description{
The \code{preMMP_results} dataset is in a long format and can be used to view the results
by party at each general election between 1890 and 1993. For an example of how to use this
dataset, refer to the Summary article found here:
\url{https://sarahcgall.github.io/scgElectionsNZ/articles/summary}.
}
\keyword{datasets}
