% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{split_electorate}
\alias{split_electorate}
\title{Split Electorate (2005 - 2023)}
\format{
A data frame with 81,269 rows and 6 variables:
\describe{
\item{Election}{election year (2005--2023)}
\item{Electorate}{electorate name (e.g., Auckland Central)}
\item{List_Party}{party voted for in party vote (e.g., 99 MP Party)}
\item{Electorate_Party}{party voted for in candidate vote (e.g., ACT Party)}
\item{Votes}{estimated number of votes flowing from list to electorate parties (n)}
\item{Percentage}{percentage of votes flowing from list to electorate parties (\%)}
}
}
\source{
This dataset was sourced from the NZ Electoral Commission's official results.
The raw data are available from \url{https://www.electionresults.govt.nz/}.
}
\usage{
split_electorate
}
\description{
The \code{split_electorate} dataset is in a long format and can be used to view the electorate-level
split voting by party. For an example of how to use this dataset, refer
to the \href{https://sarahcgall.github.io/scgElectionsNZ/articles/split-voting}{Split Voting article}.
}
\keyword{datasets}
