% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{split_total}
\alias{split_total}
\title{Split Total (2005 - 2023)}
\format{
A data frame with 2,435 rows and 5 variables:
\describe{
\item{Election}{election year (2005--2023)}
\item{List_Party}{party voted for in party vote (e.g., 99 MP Party)}
\item{Electorate_Party}{party voted for in candidate vote (e.g., ACT Party)}
\item{Votes}{number of votes flowing from list to electorate parties (n)}
\item{Percentage}{percentage of votes flowing from list to electorate parties (\%)}
}
}
\source{
This dataset was sourced from the NZ Electoral Commission's official results.
The raw data are available from \url{https://www.electionresults.govt.nz/}.
}
\usage{
split_total
}
\description{
The \code{split_total} dataset is in a long format and can be used to view the overall
split voting by party. For an example of how to use this dataset, refer
to the Split Voting article found here:
\url{https://sarahcgall.github.io/scgElectionsNZ/articles/split-voting}.
}
\keyword{datasets}
