% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_total.R
\docType{data}
\name{split_total}
\alias{split_total}
\title{Split Total (2005 - 2023)}
\format{
\subsection{\code{split_total}}{

A data frame with 2,435 rows and 5 columns:
\describe{
\item{Election}{Election year}
\item{List_Party}{Party voted for in party vote}
\item{Electorate_Party}{Party voted for in electorate vote}
\item{Votes}{Number of votes for list and electorate parties (n)}
\item{Percentage}{Percentage of votes for list and electorate parties (\%)}
@examples
split_total
}
}
}
\source{
\url{https://www.electionresults.govt.nz/}
}
\usage{
split_total
}
\description{
View overall split voting by party
}
\keyword{datasets}
