% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{data}
\name{summary}
\alias{summary}
\title{Summary (1996 - 2023)}
\format{
\subsection{\code{summary}}{

A data frame with 911 rows and 9 columns:
\describe{
\item{Election}{Election year}
\item{Ballot}{Ballot: Party Vote, Candidate Vote, or Total (Party + Candidate)}
\item{Party}{Party}
\item{Seats}{Number of seats won by party}
\item{Votes}{Vote (n)}
\item{Percentage}{Vote (\%)}
\item{Nominated}{Number nominated by party}
\item{Registered}{Yes/No: if party is registered}
\item{Successful}{Yes/No: if party achieves eligibility}
@examples
summary
}
}
}
\source{
\url{https://www.electionresults.govt.nz/}
}
\usage{
summary
}
\description{
View overall results summary data.
}
\keyword{datasets}
