% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{summary}
\alias{summary}
\title{Summary (1996 - 2023)}
\format{
A data frame with 909 rows and 9 variables:
\describe{
\item{Election}{election year (1996--2023)}
\item{Ballot}{ballot type (Candidate, Party, Total (Party + Candidate))}
\item{Party}{party name (e.g., ACT Party)}
\item{Seats}{number of seats won by party (0--65)}
\item{Votes}{total valid votes for the party (0--1.44355e+06)}
\item{Percentage}{percentage of total valid votes for the party (0.00--50.01)}
\item{Nominated}{number nominated by party on lists or ballot papers (0--153)}
\item{Registered}{if party is registered (Yes, No)}
\item{Successful}{if party achieved the eligibility requirement (Yes, No)}
}
}
\source{
This dataset was sourced from the NZ Electoral Commission's official results.
The raw data are available from \url{https://www.electionresults.govt.nz/}.
}
\usage{
summary
}
\description{
The \code{summary} dataset is in a long format and can be used to view overall
results summary data. For an example of how to use this dataset, refer
to the \href{https://sarahcgall.github.io/scgElectionsNZ/articles/summary}{Summary article}.
}
\keyword{datasets}
