% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_EName.R
\name{update_EName}
\alias{update_EName}
\title{Update Electorate Names}
\usage{
update_EName(..., column = Electorate)
}
\arguments{
\item{...}{Name of dataset}

\item{column}{Name of column with electorates}
}
\value{
A data frame with an additional region column.
}
\description{
Comvert previous electorate names to the 2020 boundary names.
}
\examples{
add_region(majority)
}
