% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vote_type}
\alias{vote_type}
\title{Vote Type (1996 - 2023)}
\format{
A data frame with 11,136 rows and 6 variables:
\describe{
\item{Election}{election year (1996--2023)}
\item{Ballot}{ballot type (Candidate, Party)}
\item{Electorate}{electorate name (e.g., Auckland Central)}
\item{Validity}{validity of the vote (Valid, Informal, Disallowed)}
\item{Vote_Type}{type of vote (Ordinary, Special - NZ, Special - Overseas)}
\item{Votes}{total number of votes cast (0--42,496)}
}
}
\source{
This dataset was sourced from the NZ Electoral Commission's official results.
The raw data are available from \url{https://www.electionresults.govt.nz/}.
}
\usage{
vote_type
}
\description{
The \code{vote_type} dataset is in a long format and can be used to view the vote type
(ordinary, special - nz, special - oversease) by validity (informal/valid/disallowed)
and Electorate and Ballot type. For an example of how to use this dataset, refer
to the \href{https://sarahcgall.github.io/scgElectionsNZ/articles/vote-type}{Vote Type article}.
}
\keyword{datasets}
