% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_type.R
\docType{data}
\name{vote_type}
\alias{vote_type}
\title{Vote Type (1996 - 2023)}
\format{
\subsection{\code{vote_type}}{

A data frame with 11,136 rows and 6 columns:
\describe{
\item{Election}{Election year}
\item{Ballot}{Ballot: Party Vote, Candidate Vote, or Total (Party + Candidate)}
\item{Electorate}{Name of Electorate}
\item{Validity}{Validity: Valid, Informal, or Disallowed}
\item{Vote_Type}{Vote_Type: Ordinary, Special - NZ, or Special - Overseas}
\item{Votes}{Votes (n)}
@examples
vote_type
}
}
}
\source{
\url{https://www.electionresults.govt.nz/}
}
\usage{
vote_type
}
\description{
View the vote type (ordinary, special - nz, special - oversease) by validity (informal/valid/disallowed)
and Electorate and Ballot type
}
\keyword{datasets}
